add authentication call in requests

+++ Backend/server.js
@@ -80,21 +80,21 @@ app.get("/api/listings/:pid", queryProperty, queryDB, (req,res)=>{
     })
 })
 
-app.put("/api/listing/update/:pid", bodyParser, updateProperty, queryDB, (req,res)=>{ // should be POST
+app.put("/api/listing/update/:pid", ensureLoggedIn, bodyParser, updateProperty, queryDB, (req,res)=>{ // should be POST
     pool.query(req.sql.getSQL(), req.sql.getValues(), function(err, data, fields) {
         if (err) throw err;
         res.json({data})
     })
 })
 
-app.post("/api/listing/create", bodyParser, createProperty, queryDB, (req,res)=>{ // should be POST
+app.post("/api/listing/create", ensureLoggedIn, bodyParser, createProperty, queryDB, (req,res)=>{ // should be POST
     pool.query(req.sql.getSQL(), req.sql.getValues(), function(err, data, fields) {
         if (err) throw err;
         res.json({data})
     })
 })
 
-app.delete("/api/listing/delete/:pid", bodyParser, deleteProperty, queryDB, (req,res)=>{ // should be POST
+app.delete("/api/listing/delete/:pid", ensureLoggedIn, bodyParser, deleteProperty, queryDB, (req,res)=>{ // should be POST
     pool.query(req.sql.getSQL(), req.sql.getValues(), function(err, data, fields) {
         if (err) throw err;
         res.json({data})
@@ -116,14 +116,14 @@ app.get("/api/accounts/:lid", queryAccount, queryDB, (req,res)=>{
     })
 })
 
-app.put("/api/accounts/update/:lid", bodyParser, updateAccount, queryDB, (req,res)=>{ // should be POST
+app.put("/api/accounts/update/:lid", ensureLoggedIn, bodyParser, updateAccount, queryDB, (req,res)=>{ // should be POST
     pool.query(req.sql.getSQL(), req.sql.getValues(), function(err, data, fields) {
         if (err) throw err;
         res.json({data})
     })
 })
 
-app.delete("/api/accounts/delete/:lid", deleteAccount, queryDB, (req,res)=>{ // should be POST
+app.delete("/api/accounts/delete/:lid", ensureLoggedIn, deleteAccount, queryDB, (req,res)=>{ // should be POST
     pool.query(req.sql.getSQL(), req.sql.getValues(), function(err, data, fields) {
         if (err) throw err;
         res.json({data})
@@ -142,7 +142,7 @@ function queryProperty(req, res, next){
 }
 
 function updateProperty(req, res, next){
-    req.body.lid=12;//change to req.user.id 
+    req.body.lid=req.session.passport.user // logged in user id
     req.body.pid=req.params.pid
     console.log(req.body)
     req.sql = new UpdatePropertyQuery(req.body);
@@ -150,13 +150,13 @@ function updateProperty(req, res, next){
 }
 
 function createProperty(req, res, next){
-    req.body.lid=12;//change to req.user.id;
+    req.body.lid=req.session.passport.user // logged in user id
     req.sql = new CreatePropertyQuery(req.body);
     next();
 }
 
 function deleteProperty(req, res, next){
-    req.body.lid=12;//change to req.user.id;
+    req.body.lid=req.session.passport.user // logged in user id
     req.body.pid=req.params.pid;
     req.sql = new DeletePropertyQuery(req.body);
     next();
@@ -169,11 +169,10 @@ function queryAccount(req, res, next){
 }
 
 function updateAccount(req, res, next){
-    // if (!req.query.lid) // default to logged-in user
+    // if (!req.params.lid) // default to logged-in user
     //     req.params.lid=req.user.id;
-    // if (req.params.lid!=req.user.id && req.user.id!=ADMIN)     // requested change doesnt match logged in user
-    //     return res.status(400).json({ err: "Bad Request"});
-    // replace with ^^ once authentication is fixed
+    if (req.params.lid!=req.session.passport.user && req.session.passport.user!=ADMIN)     // requested change doesnt match logged in user
+        return res.status(400).json({ err: "Bad Request"});
     req.body.lid=req.params.lid;
     req.sql = new UpdateAccountQuery(req.body);
     next();
@@ -182,9 +181,8 @@ function updateAccount(req, res, next){
 function deleteAccount(req, res, next){
     // if (!req.query.lid) // default to logged-in user
     //     req.params.lid=req.user.id;
-    // if (req.params.lid!=req.user.id && req.user.id!=ADMIN)     // requested change doesnt match logged in user or not admin
-    //     return res.status(400).json({ err: "Bad Request"});
-    // replace with ^^ once authentication is fixed;
+    if (req.params.lid!=req.session.passport.user && req.session.passport.user!=ADMIN)     // requested change doesnt match logged in user or not admin
+        return res.status(400).json({ err: "Bad Request"});
     req.sql = new DeleteAccountQuery(req.params);
     next();
 }
